improve SC lifetime handling
This change properly handles corner cases during SC lifetime.
To achieve this, we extend the SC lifetime beyond its corresponding cap
from userspace. When a SC cap goes out of scope (unmapped) while the
underlying SC is still attached to some thread(s), we postpone the
automatic deletion. In this case, the SC exists as a shallow version
that needs to be deleted alongside the last thread that it is attached
to. In this state, no new references to the SC can be created/attached.